wip preview-doi
